Update to 0.5.0
Added comprehensive unit tests. Literally complete code coverage, at least in terms of where it is right now, and what tests are being performed. Can't claim that it's perfect. That's not what I'm saying.

Added IlmPhase and IlmStep classes for waiting for a phase change to occur or a phase step to complete.

Lots of pytest fun.