Move prettystr to utils, rename empty_check
The `empty_check` method is internal only, so it has been renamed to `_ensure_not_none`
All files that reference it have been updated.

Also, `prettystr` is not used by all child classes, so it has been moved to `utils` as a utility function.
All files that reference it have been updated accordingly.