Release prep for 0.11.0
A ton of work went into adding unit tests for everything. Found a few glitches, I think.

This was extensively tested with es-testbed and the es-pii-tool modules before release. This is why there was no release of 0.10.0 to PyPI...

Should be more robust, and offers good defaults for pause and timeout values for everything in defaults.py